refactor: split settings-view.tsx and browser-view.tsx
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.

settings-view.tsx: already section-comment-delineated into generic form
primitives + 5 independent, prop-less section components. Split into
settings-controls.tsx (SettingRow/SectionHeader/ToggleControl/SliderControl/
KeyValueEditor) + sections/{appearance,window-panes,terminal,sounds,
advanced}-section.tsx. 812 -> 129 lines.

browser-view.tsx: split into use-pane-rect-sync.ts, use-freeze-frame.ts
(Linux-only workaround, depends on rect-sync's output), use-browser-auth.ts,
and browser-nav-bar.tsx. Hook construction order preserved in
BrowserViewInner (rect-sync before freeze-frame, since freeze-frame reads
paneRect()/paneCreated() from rect-sync's return value); each hook registers
its own onMount/onCleanup to preserve original listener-registration order,
including the auth-cancel-on-unmount safety net's timing. 797 -> 197 lines.

tsc --noEmit clean; browser-model.test.ts (16 tests, model layer, unaffected
by the view split) passes unchanged.

<!-- agentmux:agent_id=agentx -->
